Team as a single entity (#8)
Make team a single entity

This makes everyone in the team equal by removing the subdivision
or categories between the team members.
Furthermore, it makes the height of each people's card uniform
so that it looks better on each line. I find that the final result is
not only more compact, but cleaner