Adds a teaching section with the intro to numerical modeling course (#9)
This PR adds a good chunk of the GCH8108E class to the teaching section of the website. The hope of this section is to share our course content on deal.II with other people around the world that might need to use them.
Furthermore, I got tired of having to share the moodle website with people when they wanted to have the content of the course so might as well make it public :)!